refactor: PostList 컴포넌트와 PostModal 컴포넌트 분리
